test: move dhis-service-analytics integration tests into dhis-test-integration
