+++ dhis-2/dhis-services/dhis-service-analytics/pom.xml
@@ -13,31 +13,7 @@
   <packaging>jar</packaging>
   <name>DHIS Analytics</name>
 
-  <profiles>
-    <profile>
-      <id>excludeAnalyticsTests</id>
-      <activation>
-        <property>
-          <name>excludeAnalyticsTests</name>
-          <value>true</value>
-        </property>
-      </activation>
-      <build>
-        <plugins>
-          <plugin>
-            <groupId>org.apache.maven.plugins</groupId>
-            <artifactId>maven-surefire-plugin</artifactId>
-            <configuration>
-              <skipTests>true</skipTests>
-            </configuration>
-          </plugin>
-        </plugins>
-      </build>
-    </profile>
-  </profiles>
-
   <dependencies>
-
     <!-- DHIS -->
 
     <dependency>
